Extract script execution to helper function
Move the execution of external scripts out of notification.c and into
utils.c.  This will make it easier to call scripts in other situations
while retaining a consistent information-passing protocol.

A sync/async flag was in preparation of future uses of this function
relying on the return value or output of a script.