Add "hide_if" rule to conditionally hide notifications
This adds a rule to hide a notification based on the return value of a
script.  Standard UNIX convention is followed with 0 indicating "success" -
i.e. hide this notification - and non-zero indicating that a notification
should not be hidden.  If both "hide_if" and "skip_display" are given,
"hide_if" takes precedence.

Since this rule relies on a user-provided script, there is a chance that
this will hang.  Future commits will introduce a timeout on the waitpid
call.